+++ soh/soh/Enhancements/Items/InjectItemCounts.cpp
@@ -26,7 +26,7 @@ void BuildSkulltulaMessage(uint16_t* textId, bool* loadFromMessageTable) {
         msg = msg + "\x0E\x3C";
     }
     int16_t gsCount = gSaveContext.inventory.gsTokens;
-    msg.InsertNumber(gsCount);
+    msg.InsertNumber(static_cast<uint8_t>(gsCount));
     msg.AutoFormat(ITEM_SKULL_TOKEN);
     msg.LoadIntoFont();
     *loadFromMessageTable = false;
