+++ soh/soh/Enhancements/randomizer/3drando/hints.cpp
@@ -503,7 +503,7 @@ static int32_t getRandomWeight(uint32_t totalWeight) {
 
 static void DistributeAndPlaceHints(std::vector<HintDistributionSetting>& distTable, size_t totalStones) {
     auto ctx = Rando::Context::GetInstance();
-    const uint8_t junkIdx = distTable.size();
+    const uint8_t junkIdx = static_cast<uint8_t>(distTable.size());
 
     // Apply fixed hints upfront (they don't participate in weighted selection)
     for (size_t i = 0; i < distTable.size(); i++) {
